bringing exposure of globalping interface in line with codebase
