+++ server/src/controllers/settingsController.ts
@@ -1,9 +1,8 @@
 import { Request, Response, NextFunction } from "express";
-import type { IGlobalPingService } from "@/types/globalping.js";
 import { updateAppSettingsBodyValidation } from "@/validation/settingsValidation.js";
 import { sendTestEmailBodyValidation } from "@/validation/notificationValidation.js";
 import { AppError } from "@/utils/AppError.js";
-import { IEmailService, ISettingsService } from "@/service/index.js";
+import { IEmailService, ISettingsService, IGlobalPingService } from "@/service/index.js";
 
 const SERVICE_NAME = "SettingsController";
